+++ examples/api300/c7000/enclosure.rb
@@ -44,5 +44,33 @@ puts "\nUpdated #{type} '#{item[:name]}' sucessfully.\n  uri = '#{item[:uri]}'"
 # Patch update
 item.patch('replace', '/name', 'Edited_Enclosure')
 
+scope_1 = OneviewSDK::API300::C7000::Scope.new(@client, name: 'Scope 1')
+scope_1.create
+scope_2 = OneviewSDK::API300::C7000::Scope.new(@client, name: 'Scope 2')
+scope_2.create
+
+puts "\nAdding scopes"
+encl1.add_scope(scope_1)
+encl1.refresh
+puts 'Scopes:', encl1['scopeUris']
+
+puts "\nReplacing scopes"
+encl1.replace_scopes(scope_2)
+encl1.refresh
+puts 'Scopes:', encl1['scopeUris']
+
+puts "\nRemoving scopes"
+encl1.remove_scope(scope_1)
+encl1.remove_scope(scope_2)
+encl1.refresh
+puts 'Scopes:', encl1['scopeUris']
+
+scope_1.refresh
+scope_2.refresh
+
+# Delete scopes
+scope_1.delete
+scope_2.delete
+
 item.remove
 puts "\nSucessfully removed #{type} '#{item[:name]}'."