+++ spec/integration/resource/api300/c7000/scope/delete_spec.rb
@@ -11,16 +11,17 @@
 
 require 'spec_helper'
 
-klass = OneviewSDK::ImageStreamer::API300::PlanScripts
-RSpec.describe klass, integration_i3s: true, type: DELETE do
-  include_context 'integration i3s api300 context'
+klass = OneviewSDK::API300::C7000::Scope
+RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
+  include_context 'integration api300 context'
 
   describe '#delete' do
-    it 'removes a plan script' do
-      item = klass.find_by($client_i3s_300, name: PLAN_SCRIPT1_NAME_UPDATE).first
-      expect(item['uri']).to be
-      expect { item.delete }.not_to raise_error
-      expect(item.retrieve!).to eq(false)
+    it 'should delete scope' do
+      items = klass.get_all($client_300)
+      items.each do |item|
+        expect { item.delete }.not_to raise_error
+        expect(item.retrieve!).to eq(false)
+      end
     end
   end
 end